backend/requirements.txt
django==4.0.*
djangorestframework==3.13.*
django-environ==0.8.*
django_filter=21.1.*